Support YouTube short links (youtu.be) in sermon video player
Extract video URL parsing into shared utility (videoUrl.js) so both
the sermon template and OEmbed component use the same logic. This fixes
youtu.be short links not being recognized in sermon pages.

- Create src/components/shared/videoUrl.js with getYouTubeEmbedUrl and
  getVimeoEmbedUrl
- Update oembed.js to import from shared utility
- Rewrite getSermonVideoPlayer in sermon.js to use shared utility

Fixes #476
